% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intr.R
\name{intervalize}
\alias{intervalize}
\title{Expand a time-to-event data to a binned time data}
\usage{
intervalize(data, id, first, last, length)
}
\arguments{
\item{data}{a time-to-event data}

\item{id}{the name of an ID variable}

\item{first}{the name of a variable for when a participant enters the cohort}

\item{last}{the name of a variable for when a participant exits the cohort}

\item{length}{an integer for the length of a bin}
}
\value{

}
\description{
Expand a time-to-event data to a binned time data
}
